Mac ESF: Harden ESF integration for production readiness
- Switch from AUTH_OPEN to NOTIFY_OPEN to avoid ESF deadline SIGKILL
- Add inverted target path muting (kernel-level OnAccessIncludePath filtering)
- Add dedup cache (4096 slots, 3s cooldown) to suppress duplicate events
- Add NULL guards for ESF nullable pointers (executable, file)
- Guard fanotify-specific ExcludeUID check with HAVE_SYS_FANOTIFY_H
- Check es_unmute_all_target_paths and es_respond_auth_result return values
- Fix retry exhaustion silently clearing error (now logs and preserves error)
- Widen bool_opts to uint16_t for ONAS_SCTH_B_ESF flag
- Fix log message AUTH_OPEN -> NOTIFY_OPEN
- Remove duplicate b_inotify assignment
- Guard fan_fd assignment with HAVE_SYS_FANOTIFY_H
- Add build-and-start.sh script and ESF entitlements

+++ clamonacc/scan/thread.h
@@ -68,7 +68,7 @@ struct onas_scan_event {
     int32_t scantype;
     int64_t maxstream;
     int64_t timeout;
-    uint8_t bool_opts;
+    uint16_t bool_opts;
 } __attribute((packed));
 
 #ifdef HAVE_PRAGMA_PACK
